requirements.txt
# LLM Safety Gateway Dependencies

# FastAPI & Server
fastapi
uvicorn[standard]
pydantic

# Environment Variables
python-dotenv

# ML & NLP
transformers-interpret
transformers
torch
sentence-transformers
nltk
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85

# Text Processing
rapidfuzz

# Vector Database
chromadb

# LLM API (for Sanitizer)
openai

# Data Processing
pandas
datasets
evaluate
scikit-learn
